chore(core): move records app to dev showcase
